feat: Added ECDSA PKey validation and unit test
